splitted logic for better testing
